`digisossak` package med under-pakkene `forelopigsvar`, `hendelser`, `saksstatus` og `soknadsstatus`
