+++ src/main/kotlin/no/nav/sosialhjelp/innsyn/digisossak/forelopigsvar/ForelopigSvarResponse.kt
@@ -1,0 +1,6 @@
+package no.nav.sosialhjelp.innsyn.digisossak.forelopigsvar
+
+data class ForelopigSvarResponse(
+    val harMottattForelopigSvar: Boolean,
+    val link: String?
+)
