+++ src/main/kotlin/no/nav/sosialhjelp/innsyn/digisossak/hendelser/HendelseResponse.kt
@@ -1,0 +1,9 @@
+package no.nav.sosialhjelp.innsyn.digisossak.hendelser
+
+import no.nav.sosialhjelp.innsyn.domain.UrlResponse
+
+data class HendelseResponse(
+    val tidspunkt: String,
+    val beskrivelse: String,
+    val filUrl: UrlResponse?
+)
