+++ src/main/kotlin/no/nav/sosialhjelp/innsyn/digisossak/saksstatus/SaksStatusController.kt
@@ -1,12 +1,11 @@
-package no.nav.sosialhjelp.innsyn.rest
+package no.nav.sosialhjelp.innsyn.digisossak.saksstatus
 
 import no.nav.security.token.support.core.api.ProtectedWithClaims
-import no.nav.sosialhjelp.innsyn.domain.ForelopigSvarResponse
-import no.nav.sosialhjelp.innsyn.service.forelopigsvar.ForelopigSvarService
 import no.nav.sosialhjelp.innsyn.tilgang.Tilgangskontroll
 import no.nav.sosialhjelp.innsyn.utils.IntegrationUtils.ACR_LEVEL4
 import no.nav.sosialhjelp.innsyn.utils.IntegrationUtils.SELVBETJENING
-import org.springframework.http.HttpHeaders
+import org.springframework.http.HttpHeaders.AUTHORIZATION
+import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PathVariable
@@ -17,16 +16,19 @@ import org.springframework.web.bind.annotation.RestController
 @ProtectedWithClaims(issuer = SELVBETJENING, claimMap = [ACR_LEVEL4])
 @RestController
 @RequestMapping("/api/v1/innsyn")
-class ForelopigSvarController(
-    private val forelopigSvarService: ForelopigSvarService,
+class SaksStatusController(
+    private val saksStatusService: SaksStatusService,
     private val tilgangskontroll: Tilgangskontroll
 ) {
-    @GetMapping("/{fiksDigisosId}/forelopigSvar")
-    fun hentForelopigSvarStatus(@PathVariable fiksDigisosId: String, @RequestHeader(value = HttpHeaders.AUTHORIZATION) token: String): ResponseEntity<ForelopigSvarResponse> {
+
+    @GetMapping("/{fiksDigisosId}/saksStatus", produces = ["application/json;charset=UTF-8"])
+    fun hentSaksStatuser(@PathVariable fiksDigisosId: String, @RequestHeader(value = AUTHORIZATION) token: String): ResponseEntity<List<SaksStatusResponse>> {
         tilgangskontroll.sjekkTilgang(token)
 
-        val forelopigSvarResponse: ForelopigSvarResponse = forelopigSvarService.hentForelopigSvar(fiksDigisosId, token)
-
-        return ResponseEntity.ok().body(forelopigSvarResponse)
+        val saksStatuser = saksStatusService.hentSaksStatuser(fiksDigisosId, token)
+        if (saksStatuser.isEmpty()) {
+            return ResponseEntity(HttpStatus.NO_CONTENT)
+        }
+        return ResponseEntity.ok(saksStatuser)
     }
 }
