+++ src/main/kotlin/no/nav/sosialhjelp/innsyn/digisossak/soknadsstatus/SoknadsStatusController.kt
@@ -1,9 +1,7 @@
-package no.nav.sosialhjelp.innsyn.rest
+package no.nav.sosialhjelp.innsyn.digisossak.soknadsstatus
 
 import no.nav.security.token.support.core.api.ProtectedWithClaims
 import no.nav.sosialhjelp.innsyn.app.xsrf.XsrfGenerator
-import no.nav.sosialhjelp.innsyn.domain.SoknadsStatusResponse
-import no.nav.sosialhjelp.innsyn.service.soknadsstatus.SoknadsStatusService
 import no.nav.sosialhjelp.innsyn.tilgang.Tilgangskontroll
 import no.nav.sosialhjelp.innsyn.utils.IntegrationUtils.ACR_LEVEL4
 import no.nav.sosialhjelp.innsyn.utils.IntegrationUtils.SELVBETJENING
