+++ src/main/kotlin/no/nav/sosialhjelp/innsyn/rest/SaksOversiktController.kt
@@ -15,6 +15,7 @@ import no.nav.sosialhjelp.innsyn.client.dialog.DialogClient
 import no.nav.sosialhjelp.innsyn.client.dialog.DialogException
 import no.nav.sosialhjelp.innsyn.client.dialog.DialogStatus
 import no.nav.sosialhjelp.innsyn.client.fiks.FiksClient
+import no.nav.sosialhjelp.innsyn.digisossak.saksstatus.SaksStatusService
 import no.nav.sosialhjelp.innsyn.domain.InternalDigisosSoker
 import no.nav.sosialhjelp.innsyn.domain.SaksDetaljerResponse
 import no.nav.sosialhjelp.innsyn.domain.SaksListeResponse
@@ -22,7 +23,6 @@ import no.nav.sosialhjelp.innsyn.domain.SaksStatus
 import no.nav.sosialhjelp.innsyn.domain.SoknadsStatus
 import no.nav.sosialhjelp.innsyn.event.EventService
 import no.nav.sosialhjelp.innsyn.service.oppgave.OppgaveService
-import no.nav.sosialhjelp.innsyn.service.saksstatus.DEFAULT_TITTEL
 import no.nav.sosialhjelp.innsyn.tilgang.Tilgangskontroll
 import no.nav.sosialhjelp.innsyn.utils.IntegrationUtils
 import no.nav.sosialhjelp.innsyn.utils.IntegrationUtils.ACR_LEVEL4
@@ -172,7 +172,7 @@ class SaksOversiktController(
 
     private fun hentNavn(model: InternalDigisosSoker): String {
         return model.saker.filter { SaksStatus.FEILREGISTRERT != it.saksStatus }.joinToString {
-            it.tittel ?: DEFAULT_TITTEL
+            it.tittel ?: SaksStatusService.DEFAULT_SAK_TITTEL
         }
     }
