+++ src/test/kotlin/no/nav/sosialhjelp/innsyn/digisossak/hendelser/HendelseServiceTest.kt
@@ -1,4 +1,4 @@
-package no.nav.sosialhjelp.innsyn.service.hendelse
+package no.nav.sosialhjelp.innsyn.digisossak.hendelser
 
 import io.mockk.clearAllMocks
 import io.mockk.every
